NOT STABLE
got started on the ability for users to download plugins from within Convergia. the user chooses
the menu option Get New Tools, and a JEditorPane opens in a dialog with the url http://trivergia.com/convergiaptl.jsp in it. this is about as far as I have come so far.

git-svn-id: svn://svn.code.sf.net/p/opengroove/code@25 124d5c61-e844-0410-bd6f-bad94b02dd72