Wrote a very simple todo list tool. This tool allows you to add things to do, and then check them off as they are done. This tool is akin to the todo list too included with most PDAs. Currently, the tool does not synchronize it's data accross the workspace, so it's more like just a local application. Expect synchronization to come soon.
Note: this tool has not been posted on the PTL yet, as I think that there is significant work to go. 

Also note: I am planning on changing the PTL so that it automatically generates the jars to download from a specific folder in the convergia sourceforge repository. this eliminates hand-building jar files. the last modified reversion number for a particular tool's folder would be used as that tool's version index.

git-svn-id: svn://svn.code.sf.net/p/opengroove/code@69 124d5c61-e844-0410-bd6f-bad94b02dd72