So I just got the most brilliant idea...
My idea is to create an annotation, and then annotate all of the timer fields in UserContext with it. That way, when it's time to start all of the timers, I don't have to have hard-coded in there a list of fields to start; it just scans for all fields annotated with the annotation. The annotation, by the way, is @TimerField.

git-svn-id: svn://svn.code.sf.net/p/opengroove/code@704 124d5c61-e844-0410-bd6f-bad94b02dd72