A lot of stuff changed, since I've been without an internet connection due to a family party...
Anyway, I made the realm server so that it works with userids as well as usernames. It won't actually communicate with anything outside of it's realm right now, it just knows about userids and how to reduce them to a username, and how to throw a FailedResponseException should it encounter a userid on a different realm.

git-svn-id: svn://svn.code.sf.net/p/opengroove/code@714 124d5c61-e844-0410-bd6f-bad94b02dd72