Thought I had changed the support email address but apparently not...
By the way, the support email address will hopefully be defunct soon, and only used as a last resort. I'm hoping to add some sort of support server, so that you fill out your question and stuff from within OpenGroove, and it sends it to the support server for the user. They then get a response as an OpenGroove message to their account, or to an email address they provide if the issue is with their account or their account messages not working. Security key issues would have to be dealt with here. Actually, the user's public keys could be sent along with the support request, so that the support server would know what keys to use to encrypt it's reply. This would have the disadvantage that a user could spoof another user's account by including their own security keys. Oh well, it was a good thought. Another idea would be to use a JEditorPane as a lightweight browser within OpenGroove that browses to something like support.opengroove.org which contains a Google Form that allows the user to submit their support request.

git-svn-id: svn://svn.code.sf.net/p/opengroove/code@720 124d5c61-e844-0410-bd6f-bad94b02dd72