Forgot to change the decryption method to synchronize on the stream.
Actually, that could have been the cause of the problem, although if it is then I haven't written OpenGrooveRealmServer correctly, as it shouldn't try to decrypt from the same stream at the same time. Anyway, it's synchronized now. I'll update OpenGrooveRealmServer, which will pull down updates via svn:externals, and see if it works then.

git-svn-id: svn://svn.code.sf.net/p/opengroove/code@761 124d5c61-e844-0410-bd6f-bad94b02dd72