Ok, a ton of modifications this time...
Changed the login frame, the workspaces tab, and the contacts tab to use SVGPanel instead of SimpleGradientPanel and FillContainer, fixed a problem with proxy storage where only the first element could be retrieved from a stored list (trying to reference any other elements returned the first element), and added some backdrops for the login frame, the contacts pane, and the new workspace pane. I also changed a border change of the X button on the taskbar notification frame to a change of insets.

UPDATE: the proxy storage changes were actually commited in the last commit but I didn't want to re-write this commit message to reflect that.

git-svn-id: svn://svn.code.sf.net/p/opengroove/code@939 124d5c61-e844-0410-bd6f-bad94b02dd72