More modifications to MessageHistoryFrame, and working on a class, UserMessageTableModel, that presents message history in a table model.
As a side note, I found a solution for the problem with Jide where it would freeze for like 20 seconds when re-parsing code. Disable the network adapter. Long story short, it's trying to search for a bunch of UNC pathnames (erroneously), which takes a few seconds (the paths don't exist anyway, they are like \\com\example\SomeClass). I sent an email to the developer of coudgarden outlining which files he needs to change to fix it.

git-svn-id: svn://svn.code.sf.net/p/opengroove/code@1312 124d5c61-e844-0410-bd6f-bad94b02dd72