I actually mentioned something like this in a previous commit, but now it's official...
OpenGroove is currently stable, and will run. This, however, will be the last commit for a couple of months where this is the case. I've started work on the peer-to-peer conversion (see a couple of commit messages back for the message that contains the link to the page describing more about it), but so far, it's just extra classes that aren't actually loaded. In the next few commits, I'm going to start changing stuff around in OpenGroove itself that will make OpenGroove not work until the whole peer-to-peer conversion is done.

git-svn-id: svn://svn.code.sf.net/p/opengroove/code@1635 124d5c61-e844-0410-bd6f-bad94b02dd72