Moved DataBlock and its implementors into their own package, since I also added on 3 additional classes/interfaces, DataBlockBuilder and its implementors, which was making the set grow large enough for its own package. I also added methods to encode and decode a Command into a DataBlock.
Added TemporaryFileStore, which allows temporary files and folders to be created.

Modified MapEngine and MapWriter to conform to the new syntax added in a previous commit.

Added the dynamics.table package, which will contain an engine that functions very similarly to Google's BigTable, in terms of the way it is used to store data. More information has also been added to desc.txt.

git-svn-id: http://opengroove.googlecode.com/svn@51 aa5ab2c2-dddc-11dd-993c-4d20d8c3c59d