Well, since I seem to be in the mood to make breaking changes...
Deleted the message-related storage classes. Messages are going to be stored on disk as serialized message objects (which are going to be moved to G4-common so that the server can deserialize them without throwing a ClassNotFoundException; I'm still thinking of writing soething like a PassThroughObject which wraps an object for serialization that it lazily deserializes, storing the data instead in a FileBlock, which would make it so that objects can be routed through the server that the server doesn't have the class for without throwing a ClassNotFoundException on deserialization) instead of in the ProxyStorage database. This might make the message history frame run a bit slower, but as I mentioned previously, a message's recipients, subject, in-reply-to field, in-reply-to subject, sender, body, and probably attachment size and count will be stored in something like a StoredMessageHeader object, and a copy of that will also be serialized to a separate folder. More notes on this change are in the googlecode repository (http://opengroove.googlecode.com/svn).

And that reminds me, some time in the future, I'm going to be moving this over to googlecode, since it's a ton faster than sf and doesn't have advertisements. So look for that commit message.

git-svn-id: svn://svn.code.sf.net/p/opengroove/code@1641 124d5c61-e844-0410-bd6f-bad94b02dd72