Added RosterMessage, a message that is intercepted by the server and used to update a user's roster. I'm also standardizing on the username "_profile" for messages that represent server actions. Ordinary usernames can't start with an underscore, so this username is guaranteed not to be used by any ordinary user.
For example, to add a new contact to your roster, you'd send a message to "opengroove.org::_profile", containing the contact to add. (assuming your account is on the opengroove.org server.) Actually, that's not quite accurate. You, as an end-user, wouldn't send a stored message to that user; OpenGroove itself sends that message for you. Although just for the heck of it (and probably more as an easter egg than anything), I could add stored message interception on the part of the server so that you could send a stored message to _profile containing the name of a contact to add, or set their local name or something like that.

git-svn-id: http://opengroove.googlecode.com/svn@129 aa5ab2c2-dddc-11dd-993c-4d20d8c3c59d