Added a limit to the maximum size of a stack trace, since I just froze my browser trying to print stack frames for an apparent infinite recursion issue in the return command. It prints out 25 frames (this will be changable in the future via a <config> option), and then prints out "...stack truncated at 25 frames, X more" where X is the number of remaining frames.
This also seems to indicate that 99 bottles of pop runs slow because appending to a text area is slow. Maybe I should change it into a vertical panel, and append labels to it as output is shown. Then I'd cut off lines after a specified value, which could be changed with <config>.

Now I'm going to go figure out what the issue with <return> is...

git-svn-id: http://opengroove.googlecode.com/svn@271 aa5ab2c2-dddc-11dd-993c-4d20d8c3c59d